S agrego una clase entidad llamada prueba y 2 tablas intermedias materia_prueba y prueba_alumno

+++ src/main/java/com/example/EscuelaPrimaria/entities/domain/Grado.java
@@ -21,7 +21,7 @@ public class Grado {
     private Long id;
 
     @Column(name = "nivel", unique = true, nullable = false)
-    private int nivel;
+    private Long nivel;
 
     @OneToOne(mappedBy = "grado")
     private Turno turno;
